// src/canisters/model/dfusion.did.d.ts
export interface EntryDigest {
  'id' : bigint,
  'title' : string,
  'creator' : Principal,
  'deleted' : boolean,
  'createAt' : Time,
  'favorites' : Array<Principal>,
  'cover' : [] | [string],
  'likes' : Array<Principal>,
  'contentDigest' : string,
}
// ...
export type Time = bigint;
export interface UserExt {
  'id' : Principal,
  'bio' : [] | [string],
  'favorites' : Array<bigint>,
  'name' : [] | [string],
  'likes' : Array<bigint>,
  'entries' : Array<bigint>,
  'followers' : Array<Principal>,
  'following' : Array<Principal>,
}